fix: handle case of symlink existing

+++ iris_pipeline/tests/test_utils.py
@@ -22,5 +22,6 @@ def get_data_from_url(filename):
         warnings.warn(f"Retrieve data for {filename} (if not cached already)")
         local_path = data.get_pkg_data_filename(filename, show_progress=True)
         local_path_suffix = local_path + ".fits"
-        os.symlink(local_path, local_path_suffix)
+        if not os.path.exists(local_path_suffix):
+            os.symlink(local_path, local_path_suffix)
     return local_path_suffix
